cli/mcp/text_to_audio/requirements.txt
mcp>=1.0.0
fastmcp>=0.1.0
elevenlabs>=1.0.0
python-dotenv>=1.0.0
